move decorator to library
